Move BodyLevelView to new file
